app/requirements.txt
streamlit>=1.41.1,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0
odfpy>=1.4.1,<2.0.0
python-dotenv>=1.0.1,<2.0.0
pandas>=2.2.3,<3.0.0
pgvector>=0.2.5,<0.3.0
requests>=2.31.0,<3.0.0
pydantic>=2.0.0,<3.0.0
anthropic>=0.8.1
playwright>=1.41.0
pytest>=7.4.3,<8.0.0
sqlalchemy>=2.0.0,<3.0.0
numpy>=1.26.0,<2.0.0
ratelimit
duckduckgo-search
fake-useragent
beautifulsoup4
selenium
webdriver-manager
thefuzz
crawl4ai
nest-asyncio
langchain
langchain-community
langchain-anthropic
langchain-cohere
sentence-transformers
transformers
torch
